# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sistema de Gerenciamento e Criação de Propostas Comerciais
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (Admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [users]
      summary: Edit user (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Deactivate user (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Admin resets a user password
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── STATIONS ───────────────────────────────────────────────────────────────
...
  /proposal-templates/{id}/use:
    post:
      operationId: useProposalTemplate
      tags: [proposalTemplates]
      summary: Create a proposal from this template
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Proposal created from template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"

  # ─── PROPOSALS ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ─── AUTH ─────────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, name, email, role, active, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ADMIN, COMERCIAL]
        active:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [ADMIN, COMERCIAL]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ADMIN, COMERCIAL]
        active:
          type: boolean
    ResetPasswordInput:
      type: object
      required: [newPassword]
      properties:
        newPassword:
          type: string

    # ─── STATIONS ─────────────────────────────────────────────────────────────
    Station:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        slogan:
          type: ["string", "null"]
        logoBase64:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Advertiser:
      type: object
      required: [id, tradeName, active, createdAt]
      properties:
        id:
          type: string
        tradeName:
          type: string
        legalName:
          type: ["string", "null"]
        cnpj:
          type: ["string", "null"]
        logoBase64:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
...
    StatBlock:
      type: object
      required: [num, suf, desc]
      properties:
        num:
          type: string
        suf:
          type: string
        desc:
          type: string
    ProposalProduct:
      type: object
      required: [id, order, qty, title, color]
      properties:
        id:
          type: string
        order:
          type: integer
        qty:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        detail:
          type: ["string", "null"]
        program:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        color:
          type: string
          enum: [BLUE, YELLOW, RED, GREEN, DARK]
    Proposal:
      type: object
      required: [id, stationId, createdById, status, propType, propMonth, propYear, stats, overlayOpacity, createdAt, updatedAt]
      properties:
        id:
          type: string
        stationId:
          type: string
        station:
          $ref: "#/components/schemas/Station"
        advertiserId:
          type: ["string", "null"]
        advertiser:
          $ref: "#/components/schemas/Advertiser"
        createdById:
          type: string
        createdBy:
          $ref: "#/components/schemas/User"
        status:
          type: string
          enum: [DRAFT, SENT, APPROVED, REJECTED, ARCHIVED]
        fromTemplateId:
          type: ["string", "null"]
        fromTemplateName:
          type: ["string", "null"]
        propType:
          type: string
        propMonth:
          type: string
        propYear:
          type: string
        campTag:
          type: ["string", "null"]
        clientLine1:
          type: ["string", "null"]
        clientLine2:
          type: ["string", "null"]
        dateStart:
          type: ["string", "null"]
        dateEnd:
          type: ["string", "null"]
        periodDesc:
          type: ["string", "null"]
        bannerBase64:
          type: ["string", "null"]
        overlayOpacity:
          type: integer
        stats:
          type: array
          items:
            $ref: "#/components/schemas/StatBlock"
        investDesc:
          type: ["string", "null"]
        investValue:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactRole:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProposalProduct"
        createdAt:
          type: string
        updatedAt:
          type: string
...